Fix compiler warnings. (#102)
* core: Use a reference for the frame_info tokens.

This avoids an unnecessary copy.


* core: Switch to unsigned int for width, height and stride variables

Fix all associated signed/unsigned mismatches as a result.


* core: Use correct qualifiers in forward declarations.

This fixes some compiler warnings.


* stages: Mark functions as override where needed.

This fixes some compiler warnings.


* core: Fix ordering in the constructor initialiser list.

This fixes some compiler warnings.


* egl_preview: Remove unused variable warnings.


* core: Fix various singed/unsigned mismatch compiler warnings.


* options: Add virtual destructor to the Options base class

This fixes a compiler warning.


* core: Fix pollfd struct initialisation.

This fixes some compiler warnings.


* postprocessor: Fix copy elision with futures by removing the std::move call.

This fixes a compiler warning.


* build: Set -Wno-deprecated-* for qt_preview.cpp

This removes a compiler warning for the qt5/QtCore/qvariant.h header.


* build: Fix compiler flags usage.

- Fix typo to enable -Wall type flags.
- Set warnings are errors.
- Set -Wno-psabi only for gcc.


* actions: Add tensorflow lite to the build tests.

+++ post_processing_stages/object_classify_tf_stage.cpp
@@ -65,7 +65,7 @@ void ObjectClassifyTfStage::readExtras(boost::property_tree::ptree const &params
 	int output = interpreter_->outputs()[0];
 	TfLiteIntArray *output_dims = interpreter_->tensor(output)->dims;
 	// Causes might include loading the wrong model, or the wrong labels file.
-	if (output_dims->data[output_dims->size - 1] != label_count_)
+	if (output_dims->data[output_dims->size - 1] != static_cast<int>(label_count_))
 		throw std::runtime_error("ObjectClassifyTfStage: Label count mismatch");
 }
